+++ lib/graphql-hive.rb
@@ -39,5 +39,10 @@ module GraphQLHive
     def stop
       GraphQLHive.configuration.usage_reporter.stop
     end
+
+    def report_schema_to_hive(schema:, options: {})
+      sdl = GraphQL::Schema::Printer.new(schema).print_schema
+      SchemaReporter.new(sdl: sdl, options: options).send_report
+    end
   end
 end