+++ spec/hive_spec.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-RSpec.describe GraphQLHive do
-  it "has a version number" do
-    expect(GraphQLHive::VERSION).not_to be nil
-  end
-end