+++ CHANGELOG.md
@@ -10,11 +10,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Works with `Schema.trace_with` you will need to update your schema to use this method. There is an example in app.rb
 - BREAKING: You must now configure the gem in an initializer. See README.md for more details.
+- Added `GraphQLHive.report_schema_to_hive` method to report your schema to Hive either on boot or in CI.
 
 ### Removed
 - BREAKING: No longer works with `Schema.use` as this is deprecated in graphql-ruby.
 - BREAKING: Namespace conflict with `GraphQL`. Everything is now named `GraphQLHive`.
 - BREAKING: Support for Ruby 3.1 as it will reach EOL in March 2025.
+- BREAKING: Removed `reporting` and `report_schema` options. You will now have to do this manually.
 
 ### Changed
 - Refactored graphql-hive.rb to us a configuration class